+++ modules/azure/vault/variables.tf
@@ -76,3 +76,10 @@ variable "vm_size" {
   type        = string
   description = "VM Size name"
 }
+
+### additional tags
+variable "additional_tags" {
+  description = "Additional tags that can be added to all resources"
+  type        = map
+  default     = {}
+}