+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow setting custom tags to all resources in Azure.
+
 ## [8.3.0] - 2022-03-29
 
 ### Changed